Fix version tagging in variables.tf when releasing
In MC version 7.1, we introduced variable description.
To enhance readibility and conform to Terraform style
guide, we add spaces around the equal sign for the default
argument of variables puppetenv_rev. This made the sed
that would set the version of MC release lapsed because
of incorrect matching whitespaces.

This commit fixes this problem by just looking for the
word "master" between quotes in variables.tf and replacing
it with the release version.