Reintroduce nouveau driver blacklisting in cloud-init
When merging puppetagent.yaml and puppetserver.yaml, we
used puppetserver.yaml as a the base canvas. Unfortunately,
the line blacklisting the nouveau drivers introduced in
commit de17c2c4c9642389a22df2737f544a0d677a3a94 was missing
from puppetserver.yaml and was not added to puppet.yaml.

When nouveau drivers are not blacklisted, kernel panics happen
when loading NVIDIA drivers. The panic can be fatal or not, but
blacklisting nouveau drivers avoid them completely.