Replace repalce os_int_network and os_int_subnet by subnet_id
We used to assume that OpenStack project would come with one
network and one subnetwork. When the assumption was false, user
could provide os_int_network or os_int_subnet or both to identify
how Magic Castle should create the ports.

However, because Magic Castle OpenStack does not create network nor
subnet, the only information really needed to create the port is the
subnet id, the network ID can be infered from the subnet. We no longer
use the name because subnet name do not have to be unique.