Ensure AWS spot instances can be persistent
Our spot instances on Amazon are not actually persistent, when they die they do not restart. It was hard to discover but I found this documented somewhere:
 ```
* For <a>RunInstances</a>, persistent Spot Instance requests are only supported when
 * <b>InstanceInterruptionBehavior</b> is set to either <code>hibernate</code> or <code>stop</code>.
 * </p>
```
By default (and in our case) `InstanceInterruptionBehavior` is set to terminate , we need to change this to `stop`. I noticed as I left a spot cluster running over the weekend, and when I came back all the compute nodes had died.