Make issuing wilcard certificate optional
The generation of the wildcard certificate is no longer necessary
as caddy handles certificate generation per vhost by itself and also
handle renewal automatically. The wildcard generation was meant to be
used when multiple proxy / jupyterhub would be used, but this has yet
proven useful. Therefore, we are making the wildcard generation an
exception instead of the general rule.