Refactor SSHFP record generation removing hardcoded algorithm names
 We loop on the hostkeys to register the corresponding SSHFP. We
thought of looping on the data sources, but this trigger an error
because Terraform is unable to predict how many elements will be
present in the map output by the key2fp datasource. While we know it
will corresponds to the number of hostkeys, Terraform does not, and
there is no way to tell it. So instead, we loop on the hostkey and
recover the algorithm index locally.

This PR also fixes issue #214.